% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMapDP.R
\name{clustMapDP}
\alias{clustMapDP}
\title{MAP-DP Clustering}
\usage{
clustMapDP(X, N0, m0, a0, c0, B0)
}
\arguments{
\item{X}{DxN matrix of data}

\item{N0}{prior count (DP concentration paramter)}

\item{m0}{cluster prior mean}

\item{a0}{cluster prior scale}

\item{c0}{cluster prior degrees of freedom}

\item{B0}{cluster prior precision (inverse covariance)}
}
\description{
It implements the algorithm descriped by
Raykov YP, Boukouvalas A, Baig F, Little MA (2016)
What to Do When K-Means Clustering Fails: A Simple yet Principled Alternative Algorithm.
PLOS ONE 11(9): e0162259. https://doi.org/10.1371/journal.pone.0162259
}
