% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustMapDP.R
\name{repmat}
\alias{repmat}
\title{Replicates the behaviour of the repmat function of MATLAB}
\usage{
repmat(X, m, n)
}
\arguments{
\item{m}{The m value for the tiling}

\item{n}{The n value for tiling}

\item{a}{The matrix to copy}
}
\description{
Replicates the behaviour of the repmat function of MATLAB
}
